Added firestore storage of stores
Firebase firestore is now implemented for storing store data, allowing registration. The rules may need to be modified, but we do not have authorization for this, so there does not appear to be any actual danger regarding it (currently anyone can write, no one can read). The firebase firestore is still accessible by us, so we can see the data. It should be hostable now, using npm start, not firebase serve. This is because of react. This may make it difficult for us to run by hosting on google firebase hosting.